Modified termination condition and removed MINEDGES check.
The termination condition was checking only for a single complete vetex in the search tree (one with a mask containing only 0's). This led to the unusual condition that when the MINEDGES check was removed, the code did significantly less work (as it apparently found a complete vertex quicker). Also, we removed the MINEDGES check in order to not prune edges too early (due to a technicality with the method of masking out edges to approximate the old style of depth).