Completely rewrote the Explore class. It now uses threads. The search depth has been changed to.
Threads: There will always be at least two threads, the main thread and a worker thread. An arbitary number of threads may be used though.

Depth: The depth is now stored in the top row of each matrix. A child of a node at depth n will have a depth of n+1 regardless of how many ones are in it's mask.