Multiple Starting Graphs
The code is changed to allow for multiple starting graphs. All base files have been updated to the a format where the graph(s) are the last thing in the file. The code to read in the graphs is a bit dodgy (it doesn't detect EOF correctly, but this is handled by exceptions thrown by the overloaded >> operator for the Matrix class). Also, the modifications to the Explorer class to handle a queue could be done better, but work well enough for the current purposes.